Move GeoTemporalAnnotation to core app
