+++ django-rgd/rgd/models/annotation.py
@@ -1,0 +1,26 @@
+from django.contrib.gis.db import models
+from django_extensions.db.models import TimeStampedModel
+
+from .common import GeospatialFeature
+
+
+class BaseAnnotation(TimeStampedModel):
+    class Meta:
+        abstract = True
+
+    caption = models.CharField(max_length=100, blank=True, null=True)
+    label = models.CharField(max_length=100, blank=True, null=True)
+    annotator = models.CharField(max_length=100, blank=True, null=True)
+    notes = models.TextField(null=True, blank=True)
+    properties = models.JSONField(null=True, blank=True)
+
+
+class GeoTemporalAnnotation(BaseAnnotation, GeospatialFeature):
+    """Annotation in world coordinates during a time frame.
+
+    This annotation could be associated with many different images/rasters.
+
+    """
+
+    start_date = models.DateTimeField(null=True, default=None, blank=True)
+    end_date = models.DateTimeField(null=True, default=None, blank=True)
